// stories/Atoms/Badge/types.tsx
export enum CoreBadgetype { 
    Primary = 'primary',
    Secondary = "secondary",
    Success = "success",
    Warning = "warning",
    Danger = "danger",
    Info = "info",
    Light = "light",
    Dark = "dark",
    OutlinePrimary = "outline-primary",
    OutlineSecondary = "outline-secondary"
}

export enum CoreBadgesize { 
    H1 = 'size-h1',
    H2 = "size-h2",
    H3 = "size-h3",
    H4 = "size-h4",
    H5 = "size-h5",
    H6 = "size-h6",
    F14 = "text-14",
    F16 = "text-16"
}