// stories/Atoms/Button/types.tsx
export enum CoreButtontype {
    Primary = "primary",
    Secondary = "secondary",
    Success = "success",
    Warning = "warning",
    Danger = "danger",
    Info = "info",
    OutlinePrimary = "outline-primary",
    OutlineSecondary = "outline-secondary",
    OutlineSuccess = "outline-success",
    OutlineWarning = "outline-warning",
    OutlineDanger = "outline-danger",
    OutlineInfo = "outline-info",
    Link = "link",
    Light = "light",
    Dark = "dark"
}